Implemented connected peer IP verification.

+++ C/p2p_cli.c
@@ -52,19 +52,19 @@ void start_cli(struct control_st *ctrl){
       printf("\t          you the name of the file you want\n");
 
     }else if(strncmp(buffer, "list", 4) == 0){
-      write(ctrl->udp_recv_pipe, "list", 4);
+      write(ctrl->udp_recv_pipe_out, "list", 4);
       if(buffer[4] != '\0'){
         err = inet_pton(AF_INET, buffer+5, &bf_addr);
         if(err == 1){
-          write(ctrl->udp_recv_pipe, " ", 1);
-          write(ctrl->udp_recv_pipe, &bf_addr.s_addr, sizeof(in_addr_t));
+          write(ctrl->udp_recv_pipe_out, " ", 1);
+          write(ctrl->udp_recv_pipe_out, &bf_addr.s_addr, sizeof(in_addr_t));
         }else if(err == 0){
           fprintf(stderr, "Invalid IPv4 address passed as an argument.\n");
         }else{
           fprintf(stderr, "Unspecified error occurred.\n");
         }
       }
-      write(ctrl->udp_recv_pipe, "\0", 1);
+      write(ctrl->udp_recv_pipe_out, "\0", 1);
       kill(ctrl->udp_recv_pid, SIGUSR1);
       sem_wait(sem);  // Wait for process to be over
 
@@ -73,27 +73,36 @@ void start_cli(struct control_st *ctrl){
       if(err == 1){
         char file[BUFFER_SIZE];
         char location[BUFFER_SIZE];
+        char temp;
 
-        write(ctrl->udp_recv_pipe, "list ", 5);
-        write(ctrl->udp_recv_pipe, &bf_addr.s_addr, sizeof(in_addr_t));
-        write(ctrl->udp_recv_pipe, "\0", 1);
+        write(ctrl->udp_recv_pipe_out, "list ", 5);
+        write(ctrl->udp_recv_pipe_out, &bf_addr.s_addr, sizeof(in_addr_t));
+        write(ctrl->udp_recv_pipe_out, "\0", 1);
         kill(ctrl->udp_recv_pid, SIGUSR1);
         sem_wait(sem);  // Wait for process to be over
 
-        printf("What file do you want to download?\n>");
-        fgets(file, BUFFER_SIZE, stdin);
-        remove_newline(file);
+        read(ctrl->udp_recv_pipe_in, &temp, sizeof(char));
+        if(temp == '0'){
+          printf("The specified peer either is not connected to the network\n"
+                "or it's not broadcasting it's presence. Cannot list files.\n");
+        }else if(temp == '1'){
+          printf("What file do you want to download?\n>");
+          fgets(file, BUFFER_SIZE, stdin);
+          remove_newline(file);
 
-        printf("Where do you want to save the file? (please insert "
-              "full path)\n>");
-        fgets(location, BUFFER_SIZE, stdin);
-        remove_newline(location);
+          printf("Where do you want to save the file? (please insert "
+                "full path)\n>");
+          fgets(location, BUFFER_SIZE, stdin);
+          remove_newline(location);
 
-        write(ctrl->udp_recv_pipe, "get ", 4);
-        write(ctrl->udp_recv_pipe, &bf_addr.s_addr, sizeof(in_addr_t));
-        write(ctrl->udp_recv_pipe, "\0", 1);
-        kill(ctrl->udp_recv_pid, SIGUSR1);
-        sem_wait(sem);  // Wait for process to be over
+          write(ctrl->udp_recv_pipe_out, "get ", 4);
+          write(ctrl->udp_recv_pipe_out, &bf_addr.s_addr, sizeof(in_addr_t));
+          write(ctrl->udp_recv_pipe_out, "\0", 1);
+          kill(ctrl->udp_recv_pid, SIGUSR1);
+          sem_wait(sem);  // Wait for process to be over
+        }else{
+          fprintf(stderr, "broadcast: Unknown response from receiver.\n");
+        }
       }else if(err == 0){
         fprintf(stderr, "Invalid IPv4 address passed as an argument.\n");
       }else{
